requirements.txt
requests==2.32.3
tqdm==4.66.4
pandas==2.2.2
python-dotenv==1.0.1
